refactor if condition
